feat(lib): return change event and expose onKeyPress callback
